Move custom PhoneNumber fields and widget to separate files
