+++ saleor/userprofile/models.py
@@ -8,11 +8,8 @@ from django.utils import timezone
 from django.utils.encoding import python_2_unicode_compatible
 from django.utils.translation import pgettext_lazy
 from django_countries.fields import Country, CountryField
-from phonenumber_field.modelfields import PhoneNumberField
 
-from .validators import validate_possible_number
-
-PhoneNumberField.default_validators = [validate_possible_number]
+from .fields import PossiblePhoneNumberModelField
 
 
 class AddressManager(models.Manager):
@@ -76,7 +73,7 @@ class Address(models.Model):
     country_area = models.CharField(
         pgettext_lazy('Address field', 'state or province'),
         max_length=128, blank=True)
-    phone = PhoneNumberField(
+    phone = PossiblePhoneNumberModelField(
         verbose_name=pgettext_lazy('Address field', 'phone number'),
         blank=True, default='')
     objects = AddressManager()
